Time: 82 ms (79.33%), Space: 47.1 MB (24.82%) - LeetHub

+++ 0560-subarray-sum-equals-k/0560-subarray-sum-equals-k.cpp
@@ -2,15 +2,14 @@ class Solution {
 public:
     int subarraySum(vector<int>& nums, int k) {
         
-        unordered_map<long long,int>mpp;
+        unordered_map<int,int>mpp;
        mpp[0]=1;
         int c=0;
-        long long sum=0;
+        int sum=0;
         for(int i=0;i<nums.size();i++)
         {
             sum+=nums[i];
-            
-            long long rem=sum-k;
+           int rem=sum-k;
             c+=mpp[rem];
             mpp[sum]+=1;
         }
